Expose an internal end point for processing of events. (#24)
This is different than event publisher end point which expects a full blown event structure. It expects a JSON Object as request which has two parts. One mandatory part is <event> this is processed as other events from the message bus. Second optional part is <context> which can provide additional information like email should be sent or not. Note that difference between original message on message bus and this is the presence of op name. Corresponding changes need to be handled in nucleus-handlers-event as well.